refactor: change code
