+++ src/pages/cart-page.ts
@@ -3,7 +3,7 @@ import ViewBuilder from '../shared/lib/view-builder';
 import ElementBuilder from '../shared/lib/element-builder';
 import CartList from '../features/cart/cart-list';
 import store from '../app/store';
-import CartApi from '../entities/cart/cart';
+import CartApi from '../entities/cart/cart-api';
 import eventBus, { EventBusActions } from '../shared/lib/event-bus';
 
 export default class CartPage extends ViewBuilder {
@@ -11,11 +11,8 @@ export default class CartPage extends ViewBuilder {
 
   constructor() {
     super('page cart-page');
-    this.loadProducts();
     eventBus.subscribe(EventBusActions.UPDATE_USER, (data) => (data ? this.loadProducts() : this.cartList.empty()));
-    eventBus.subscribe(EventBusActions.UPDATE_CART, (data) =>
-      data ? this.cartList.setCards(data as Cart) : this.cartList.empty(),
-    );
+    store.user ? this.loadProducts() : this.cartList.empty();
   }
 
   public configureView(): HTMLElement[] {
@@ -34,18 +31,16 @@ export default class CartPage extends ViewBuilder {
   }
 
   public loadProducts(): void {
-    if (store.user) {
-      if (localStorage.getItem('cartID') && localStorage.getItem('token_store')) {
-        CartApi.getCustomerCart().then((data: ClientResponse<Cart>) => {
-          store.setCart(data.body);
-          this.cartList.setCards(data.body);
-        });
-      } else if (localStorage.getItem('cartID') && !localStorage.getItem('token_store')) {
-        CartApi.getAnonymousCart().then((data: ClientResponse<Cart>) => {
-          store.setCart(data.body);
-          this.cartList.setCards(data.body);
-        });
-      }
+    if (localStorage.getItem('cartID') && localStorage.getItem('token_store')) {
+      CartApi.getCustomerCart().then((data: ClientResponse<Cart>) => {
+        store.setCart(data.body);
+        this.cartList.setCards(data.body);
+      });
+    } else if (localStorage.getItem('cartID') && !localStorage.getItem('token_store')) {
+      CartApi.getAnonymousCart().then((data: ClientResponse<Cart>) => {
+        store.setCart(data.body);
+        this.cartList.setCards(data.body);
+      });
     }
   }
 }
